backend/backend/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
linecache2==1.0.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traceback2==1.4.0
typing-extensions==3.10.0.0
